+++ keanu-project/src/main/java/io/improbable/keanu/util/io/ProtobufLoader.java
@@ -1,4 +1,4 @@
-package io.improbable.keanu.network;
+package io.improbable.keanu.util.io;
 
 import com.google.common.primitives.Booleans;
 import com.google.common.primitives.Doubles;
@@ -6,6 +6,8 @@ import com.google.common.primitives.Ints;
 import com.google.common.primitives.Longs;
 import com.google.gson.internal.Primitives;
 import io.improbable.keanu.KeanuSavedBayesNet;
+import io.improbable.keanu.network.BayesianNetwork;
+import io.improbable.keanu.network.NetworkLoader;
 import io.improbable.keanu.tensor.bool.BooleanTensor;
 import io.improbable.keanu.tensor.dbl.DoubleTensor;
 import io.improbable.keanu.tensor.intgr.IntegerTensor;
